# Feature/Optimize Node Updates [PLAT-574]