--- a/api/base/serializers.py
+++ b/api/base/serializers.py
@@ -369,6 +369,19 @@ class JSONAPIListField(ser.ListField):
         return super(JSONAPIListField, self).to_internal_value(data)
 
 
+class ValuesListField(JSONAPIListField):
+        """
+        JSONAPIListField that uses a values_list with flat=True to return just
+        an array of the specified field (attr_name) for optimization purposes.
+        """
+    def __init__(self, **kwargs):
+        self.attr_name = kwargs.pop('attr_name')
+        super(ValuesListField, self).__init__(**kwargs)
+
+    def to_representation(self, val):
+        return val.values_list(self.attr_name, flat=True)
+
+
 class AuthorizedCharField(ser.CharField):
     """
     Passes auth of the logged-in user to the object's method
@@ -1252,10 +1265,7 @@ class JSONAPISerializer(BaseAPISerializer):
             else:
                 try:
                     if hasattr(attribute, 'all'):
-                        if field.field_name == 'tags':
-                            representation = attribute.values_list('name', flat=True)
-                        else:
-                            representation = field.to_representation(attribute.all())
+                        representation = field.to_representation(attribute.all())
                     else:
                         representation = field.to_representation(attribute)
                 except SkipField: