--- a/api/nodes/permissions.py
+++ b/api/nodes/permissions.py
@@ -76,8 +76,11 @@ class AdminOrPublic(permissions.BasePermission):
 class ExcludeWithdrawals(permissions.BasePermission):
 
     def has_object_permission(self, request, view, obj):
-        context = request.parser_context['kwargs']
-        node = AbstractNode.load(context[view.node_lookup_url_kwarg])
+        if obj.__class__.__name__ == 'Node':
+            node = obj
+        else:
+            context = request.parser_context['kwargs']
+            node = AbstractNode.load(context[view.node_lookup_url_kwarg])
         if node.is_retracted:
             return False
         return True