--- a/osf/models/base.py
+++ b/osf/models/base.py
@@ -304,8 +304,8 @@ class GuidMixin(BaseIDMixin):
         try:
             # guids___id__isnull=False forces an INNER JOIN
             if select_for_update:
-                return cls.objects.filter(guids___id__isnull=False, guids___id=q).select_for_update().first()
-            return cls.objects.filter(guids___id__isnull=False, guids___id=q).first()
+                return cls.objects.filter(guids___id__isnull=False, guids___id=q).select_for_update()[:1].get()
+            return cls.objects.filter(guids___id__isnull=False, guids___id=q)[:1].get()
         except cls.DoesNotExist:
             return None