--- a/osf/models/registrations.py
+++ b/osf/models/registrations.py
@@ -357,9 +357,21 @@ class Registration(AbstractNode):
         else:
             raise NodeStateError('Cannot add tags to withdrawn registrations.')
 
+    def add_tags(self, tags, auth=None, save=True, log=True, system=False):
+        if self.retraction is None:
+            super(Registration, self).add_tags(tags, auth, save, log, system)
+        else:
+            raise NodeStateError('Cannot add tags to withdrawn registrations.')
+
     def remove_tag(self, tag, auth, save=True):
         if self.retraction is None:
             super(Registration, self).remove_tag(tag, auth, save)
+        else:
+            raise NodeStateError('Cannot remove tags of withdrawn registrations.')
+
+    def remove_tags(self, tags, auth, save=True):
+        if self.retraction is None:
+            super(Registration, self).remove_tags(tags, auth, save)
         else:
             raise NodeStateError('Cannot remove tags of withdrawn registrations.')