<!-- Before submit your Pull Request, make sure you picked
     the right branch:

     - For hotfixes, select "master" as the target branch
     - For new features, select "develop" as the target branch
     - For release feature fixes, select the relevant release branch (release/X.Y.Z) as the target branch -->

## Purpose

According the NewRelic, NodeDetail.patch has the slowest average response time.

Average: 4,900 ms
SD: 22,000 ms
Min: 31.8 ms
Max: 169,000 ms

Investigate this endpoint and see what can be done to improve it.

## Changes

Very minor optimizations added here.  Node updates cause ~130 queries on average, and a lot of those queries are necessary (for example, loading a Node by guid, get_user(), when privacy is changed, all addons have to be looked at since they are in different tables.) . The node update request is slow, but not as bad as New Relic makes it seem, look at the standard deviation above. 
 
### Before
  | AVG | STDEV
-- | -- | --
Overall, ms | 2,168. | 330.
Queries, ms | 1,174. | 245.
Num queries | 131. | 5.

### After

  | AVG | STDEV
-- | -- | --
Overall, ms | 2,107. | 173.
Queries, ms | 1,082. | 120.
Num queries | 123. | 9.

## More info:
Lots of exploration done, experimented with[ turning some of the expensive queries into sql](https://github.com/CenterForOpenScience/osf.io/commit/1aa1bbcaf81c1ce41bf6afd558d521155da3025a), but there weren't massive improvements, so not worth adding to the codebase.

The more fields that are updated, the slower the query becomes.  Testing was done with a `PATCH localhost:8000/v2/nodes/12345` with variants on the following request body, each time changing the title, description, category, adding and removing tags, and turning public => private or vice versa.
```
 {
     "data": {
       "type": "nodes",
       "id": "12345",
       "attributes": {
         "title": "Test title",
         "description": "test description",
         "category": "software",
         "tags": ["first tag", "second tag", "third tag"],
         "public": true
       }
     }
}

```


## QA Notes
API tests should cover this

## To note:
For future reference, I used [django-silk](https://github.com/jazzband/django-silk) to profile `PATCH` requests locally.  To get working locally, follow the instructions in the docs, and also uncomment ` 'django.contrib.sessions.middleware.SessionMiddleware',` and run migrations.
![screen shot 2018-02-09 at 4 29 43 pm](https://user-images.githubusercontent.com/9755598/36053224-75a59394-0db6-11e8-9dd6-8e8630f08c22.png)

## Side Effects

<!-- Any possible side effects? -->

## Ticket
https://openscience.atlassian.net/browse/PLAT-574